+++ packages/sdk-release-kit/src/changelog/scripts/update-changelog.js
@@ -9,5 +9,8 @@ const changelogContents = readFileSync(changeLogFilePath, 'utf8')
 const version = JSON.parse(readFileSync(path.resolve(targetFolder, 'package.json'), 'utf8')).version
 
 module.exports = () => {
-  writeFileSync(changeLogFilePath, updateChangelogContents({changelogContents, version}))
+  writeFileSync(
+    changeLogFilePath,
+    updateChangelogContents({changelogContents, version, withDate: true}),
+  )
 }